requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
# Векторна база даних та пошук
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Обробка документів
pypdf>=5.1.0
pdfplumber>=0.11.0

# LLM (OpenAI або локальна через Ollama)
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
tiktoken>=0.7.0

# Інтерфейс та середовище
streamlit>=1.40.0
python-dotenv>=1.0.1

# Додатково для коректної роботи в Docker (Linux)
pydantic>=2.9.0
langchain-text-splitters